refactor: global 디자인 폴더 변경 및 디렉토리 globals 하위로 이동
