+++ src/styles/globals/global.ts
